requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pinecone>=6.0.0
sentence-transformers>=3.0.0
PyPDF2>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
rank-bm25>=0.2.2
groq>=0.9.0
numpy>=1.26.0
tqdm>=4.66.0
